+++ CotEditor/Sources/AppearancePaneController.swift
@@ -115,7 +115,7 @@ final class AppearancePaneController: NSViewController, NSMenuItemValidation, NS
             self.darkAppearanceButton?.state = .on
         }
         
-        let themeName = ThemeManager.shared.userDefaultSettingName(forDark: self.view.effectiveAppearance.isDark)
+        let themeName = ThemeManager.shared.userDefaultSettingName
         let row = self.themeNames.firstIndex(of: themeName) ?? 0
         self.themeTableView?.selectRowIndexes(IndexSet(integer: row), byExtendingSelection: false)
     }
@@ -299,10 +299,13 @@ final class AppearancePaneController: NSViewController, NSMenuItemValidation, NS
         // update default theme setting
         if UserDefaults.standard[.theme] != themeName {
             // do not store to UserDefautls if it's the default theme
-            if ThemeManager.shared.defaultSettingName(forDark: self.view.effectiveAppearance.isDark) == themeName {
+            if ThemeManager.shared.defaultSettingName == themeName {
                 UserDefaults.standard.restore(key: .theme)
+                UserDefaults.standard[.pinsThemeAppearance] = false
             } else {
+                let isDarkTheme = ThemeManager.shared.isDark(name: themeName)
                 UserDefaults.standard[.theme] = themeName
+                UserDefaults.standard[.pinsThemeAppearance] = (isDarkTheme != NSAppearance.current.isDark)
             }
         }
         
@@ -534,7 +537,7 @@ final class AppearancePaneController: NSViewController, NSMenuItemValidation, NS
     @objc private dynamic var selectedThemeName: String {
         
         guard let tableView = self.themeTableView, tableView.selectedRow >= 0 else {
-            return ThemeManager.shared.userDefaultSettingName(forDark: self.view.effectiveAppearance.isDark)
+            return ThemeManager.shared.userDefaultSettingName
         }
         return self.themeNames[tableView.selectedRow]
     }
@@ -622,7 +625,7 @@ final class AppearancePaneController: NSViewController, NSMenuItemValidation, NS
     /// update theme list
     @objc private func setupThemeList() {
         
-        let themeName = ThemeManager.shared.userDefaultSettingName(forDark: self.view.effectiveAppearance.isDark)
+        let themeName = ThemeManager.shared.userDefaultSettingName
         
         self.themeNames = ThemeManager.shared.settingNames
         self.themeTableView?.reloadData()