+++ CotEditor/Sources/ThemeManager.swift
@@ -24,6 +24,7 @@
 //
 
 import Foundation
+import AppKit.NSAppearance
 
 @objc protocol ThemeHolder: AnyObject {
     
@@ -77,19 +78,28 @@ final class ThemeManager: SettingFileManaging {
     
     // MARK: Public Methods
     
-    /// return default setting by taking the appearance state into consideration
-    func defaultSettingName(forDark: Bool = false) -> String {
+    /// default setting by taking the appearance state into consideration
+    var defaultSettingName: String {
         
         let defaultSettingName = DefaultSettings.defaults[.theme] as! String
+        let forDark = self.usesDarkAppearance
         
         return self.equivalentSettingName(to: defaultSettingName, forDark: forDark)!
     }
     
     
-    /// return user default setting by taking the appearance state into consideration
-    func userDefaultSettingName(forDark: Bool = false) -> String {
+    /// user default setting by taking the appearance state into consideration
+    var userDefaultSettingName: String {
         
-        return UserDefaults.standard[.theme] ?? self.defaultSettingName(forDark: forDark)
+        let settingName = UserDefaults.standard[.theme]!
+        
+        if UserDefaults.standard[.pinsThemeAppearance] || NSAppKitVersion.current < .macOS10_13 {
+            return settingName
+        }
+        
+        let forDark = self.usesDarkAppearance
+        
+        return self.equivalentSettingName(to: settingName, forDark: forDark) ?? self.defaultSettingName
     }
     
     
@@ -133,6 +143,13 @@ final class ThemeManager: SettingFileManaging {
         try self.save(settingDictionary: self.blankSettingDictionary, name: name) {
             completionHandler(name)
         }
+    }
+    
+    
+    /// return whether given setting name is dark theme
+    func isDark(name: String) -> Bool {
+        
+        return name.range(of: "(Dark)", options: [.anchored, .backwards]) != nil
     }
     
     
@@ -190,6 +207,20 @@ final class ThemeManager: SettingFileManaging {
     
     
     // MARK: Private Methods
+    
+    /// Whether user prefers using dark mode window.
+    private var usesDarkAppearance: Bool {
+        
+        switch UserDefaults.standard[.documentAppearance] {
+        case .default:
+            return NSAppearance.current.isDark
+        case .light:
+            return false
+        case .dark:
+            return true
+        }
+    }
+        
     
     /// Load ThemeDictionary from a file at the URL.
     ///